* commit 'upstream/0.1.1':
  win32 build: default is static libs
  removed optimization flags
  fixed compilation on windows (issues with finding boost libraries)
  building gdf_merger in dynamic mode no longer trys to link GDF.lib
  fixed row/column ordering of data
  clarified exception text when there is an header issue
  added option to arrange data in rows or columns to the matlab gdf_reader.
  fixed EventConverter
  fixed a compiler warning in MSVC.
  fixed argument error checking for the matlab gdf_reader.
  added documentation of how the specification was implemented.

* debian:
  added patch up_cp-data-orientation to "cherry-pick" 85d22dfc1f3191d4b70ce9f4a76f8a66700885e5 (fixed data orientation in the resampling example)
  added patch up_cp-ev-channel to "cherry-pick" f309cc6c0ede3cf3cded1cb26fd19958d3654402 (fixed event channel numbering in gdf_writer)
  added patch up_cp-ev to "cherry-pick" 96a0595d953b17b5eb1dd9937c8bfdf0d5f5579d (fixed event mode 3 handling in the resampling example)
  changelog entry for 0.1.1-1
  BF: matlab-gdf -- corrected path to the distdir
  fixing up debian/copyright
  boosting up versioned bdepends (cmake 2.8, boost past 1.34.1)

+++ libgdf/src/GDFHeaderAccess.cpp
@@ -442,6 +442,6 @@ namespace gdf
         if( it == m_sighdr.end() )
             return 0;
 
-        return m_mainhdr.get_num_datarecords() * it->second.get_samples_per_record();
+        return boost::numeric_cast<size_t>(m_mainhdr.get_num_datarecords()) * boost::numeric_cast<size_t>(it->second.get_samples_per_record());
     }
 }
